move utility functions to lib
