+++ ckanext/report/controllers.py
@@ -1,13 +1,11 @@
-from builtins import str, int
-
-import datetime
+from builtins import str
 
 from ckan.lib.helpers import json
 import ckan.plugins.toolkit as t
 import ckanext.report.helpers as helpers
 from ckanext.report.report_registry import Report
 from ckan.lib.render import TemplateNotFound
-from collections import OrderedDict
+from ckanext.report.lib import make_csv_from_dicts, ensure_data_is_dicts, anonymise_user_names
 
 
 log = __import__('logging').getLogger(__name__)
@@ -136,76 +134,3 @@ class ReportController(t.BaseController):
             'options_html': options_html,
             'report_template': report['template'],
             'are_some_results': are_some_results})
-
-
-def make_csv_from_dicts(rows):
-    import csv
-    import cStringIO as StringIO
-
-    csvout = StringIO.StringIO()
-    csvwriter = csv.writer(
-        csvout,
-        dialect='excel',
-        quoting=csv.QUOTE_NONNUMERIC
-    )
-    # extract the headers by looking at all the rows and
-    # get a full list of the keys, retaining their ordering
-    headers_ordered = []
-    headers_set = set()
-    for row in rows:
-        new_headers = set(row.keys()) - headers_set
-        headers_set |= new_headers
-        for header in row.keys():
-            if header in new_headers:
-                headers_ordered.append(header)
-    csvwriter.writerow(headers_ordered)
-    for row in rows:
-        items = []
-        for header in headers_ordered:
-            item = row.get(header, 'no record')
-            if isinstance(item, datetime.datetime):
-                item = item.strftime('%Y-%m-%d %H:%M')
-            elif isinstance(item, (int, float, list, tuple)):
-                item = str(item)
-            elif item is None:
-                item = ''
-            else:
-                item = item.encode('utf8')
-            items.append(item)
-        try:
-            csvwriter.writerow(items)
-        except Exception as e:
-            raise Exception("%s: %s, %s" % (e, row, items))
-    csvout.seek(0)
-    return csvout.read()
-
-
-def ensure_data_is_dicts(data):
-    '''Ensure that the data is a list of dicts, rather than a list of tuples
-    with column names, as sometimes is the case. Changes it in place'''
-    if data['table'] and isinstance(data['table'][0], (list, tuple)):
-        new_data = []
-        columns = data['columns']
-        for row in data['table']:
-            new_data.append(OrderedDict(zip(columns, row)))
-        data['table'] = new_data
-        del data['columns']
-
-
-def anonymise_user_names(data, organization=None):
-    '''Ensure any columns with names in are anonymised, unless the current user
-    has privileges.
-
-    NB this is only enabled for data.gov.uk - it is custom functionality.
-    '''
-    try:
-        import ckanext.dgu.lib.helpers as dguhelpers
-    except ImportError:
-        # If this is not DGU then cannot do the anonymization
-        return
-    column_names = data['table'][0].keys() if data['table'] else []
-    for col in column_names:
-        if col.lower() in ('user', 'username', 'user name', 'author'):
-            for row in data['table']:
-                row[col] = dguhelpers.user_link_info(
-                    row[col], organization=organization)[0]
